Small change of error message for no write access user
The file has already been unpacked, so we don't need the suffix for source in the cp command. Found an already existing variable with a path to the install file, so I'm using it instead of manually composing a path. Also I'm suggesting to use mv instead of cp as the user won't need a temporary file in the directory.